+++ Demos/CEthernet/EthernetDemo.c
@@ -203,9 +203,23 @@ int main(int argc, char* argv[])
     }
 
     SilKit_ReturnCode returnCode;
-    returnCode = SilKit_Participant_Create(&participant, jsonString, participantName, registryUri, SilKit_False);
+
+    SilKit_ParticipantConfiguration* participantConfiguration = NULL;
+    returnCode = SilKit_ParticipantConfiguration_FromString(&participantConfiguration, jsonString);
+    if (returnCode) {
+        printf("%s\n", SilKit_GetLastErrorString());
+        return 2;
+    }
+
+    returnCode = SilKit_Participant_Create(&participant, participantConfiguration, participantName, registryUri);
     if (returnCode) 
     {
+        printf("%s\n", SilKit_GetLastErrorString());
+        return 2;
+    }
+
+    returnCode = SilKit_ParticipantConfiguration_Destroy(participantConfiguration);
+    if (returnCode) {
         printf("%s\n", SilKit_GetLastErrorString());
         return 2;
     }